requirements.txt
# ==========================================
# ARGUS - Self-Healing PC Monitor & Attack Simulator
# ==========================================

# Core monitoring
psutil==5.9.8
pywin32==306; sys_platform == "win32"
wmi==1.5.1; sys_platform == "win32"

# Web framework (async-first)
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.27

# AI / ML
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
shap==0.44.1
joblib==1.3.2

# Optional: LSTM autoencoder (heavy deps — install if using)
# tensorflow==2.15.0
# keras==2.15.0

# Config
pyyaml==6.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Notifications
win10toast==0.9; sys_platform == "win32"
plyer==2.1.0

# Alerts (optional channels)
python-telegram-bot==21.0
discord-webhook==1.3.0

# Utilities
colorama==0.4.6
pynput==1.7.6
requests==2.31.0
python-dotenv==1.0.1

# Testing (dev)
pytest==8.0.0
pytest-asyncio==0.23.5
hypothesis==6.98.0
